lgc: set scalarizer min-bits to 32
Do not break up packed 16-bit math.

The test changes are admittedly a bit haphazard, but their point is
primarily to focus the testing on how the frontend lowers SPIR-V to LGC,
and to make the tests independent of future decisions regarding how and
whether exactly we do scalarization.

This depends on https://reviews.llvm.org/D149842/