Add vkgc_gpurtshim library
This shim allows LLPC to talk directly to the GPURT component when it is
available, i.e. whenever we build from a source tree that includes the
driver.

Standalone builds of amdllpc will not be able to talk to GPURT.

A separate library is used primarily to prevent the gpurt include paths
from "leaking" into LLPC proper, to avoid the risk of accidentally
creating an unintended build dependency.