Mark image descriptor loads invariant
Mark image/sampler descriptor loads invariant in the SPIR-V reader
to allow more code motion optimizations in the middle-end.

This has a large impact on sgpr pressure. The total number of
v_writelane instructions in the 10k shaders set (which is a rough
indicator of excessive sgpr pressure) goes from 14702 to 8805.