llpc: remove PipelineContext::hasRayQuery()
The virtual method wasn't used consistently to check for the presence of
RayQuery, and it wasn't even implemented properly in all derived
classes.

This should mostly be NFC (no functional change), except that the wave
size behavior is now improved in the hypothetical case of a graphics
pipeline in which only one shader stage uses ray queries.

Also clarify the meaning of ShaderModuleUsage::enableRayQuery. Ideally,
we'd rename that variable to useRayQuery, but that requires coordination
with the driver.

v2:
- remove the method entirely, even from GraphicsContext