llpc/raytracing: rename the traceRayShaderInfo to representativeShaderInfo
Within the raytracing context, it really doesn't play the role of a
"traceray shader", and in any case the term is misleading here.

What this really is is a container into which we can put shader options
that represent options for the pipeline as a whole, in a way that is
understood by generic shared methods.

We keep a local copy to simplify ownership, and to reduce the risk that
we accidentally end up using per-shader information that really isn't
pipeline-wide.