vkgc/llpc: obtain the GPURT shader library from GPURT by default
Refactor the vkgc interface so that:

* all GPURT shader library information is consolidated in RtState

* the compiler has all information needed to obtain the shader library
  directly from the GPURT component

* the compiler is made aware whether the shader library was explicitly
  overridden via a panel setting

The driver is still expected to fill in the default shader library, even
if it isn't overridden by a panel setting. This serves two purposes:

* simplify the driver-side implementation as long as the BIL interface
  isn't changed in the same way

* small concession to standalone builds of amdllpc (i.e., builds of
  amdllpc from a source tree that doesn't include the driver)

The primary motivation of this change, other than cleaning up code
duplication, is to align the vkgc interface with the desire to make the
command-line amdllpc compiler more useful:

* increase the stability of raytracing pipeline dumps, i.e. raytracing
  pipeline dumps no longer become stale as easily due to drift in the
  GPURT interface

* make raytracing .pipe lit tests feasible (this is a combination of the
  previous bullet and the fact that we don't want to add the GPURT
  shader library SPIR-V into the llpc/test subdirectory)

* allow portability of raytracing pipeline dumps between different gfxip
  versions

* allow amdllpc to compile individual GLSL or SPIR-V shaders that use
  raytracing features (ray query or RT pipeline shaders)

This commit enables all but the last bullet.

Note that this change also makes it easier to test a gpurt change across
a database of collected pipeline dumps, by either recompiling amdllpc or
by using the -gpurt-library command-line option to override the default.

v2:
- review comments and formatting
- add a missing fclose()