Add rasterStream to vkgcDefs.h
The rasterization stream ID is supposed to be set by driver. This is
not simply checked by detecting the stream qualifier for GS built-in
output. Rather, this can be missing.

In this change, we add the rasterStream to the member of rsState. In the
future, driver will pass this info to compiler.