Move module linking from before SpirvLower to after
This module linking position movement is required for the second stage of
refactoring, when LowerRayTracing is at the end of the SpirvLower
passes aka
AlwaysInlinerPass();
SpirvLowerGlobal();
LowerRayTracing();

when all the functions are inlined from combinded/linking modules if
module linking before spirvLower. and
there is no functions left to lower.

Without this step, the first stage of refactoring will collect and
analyse the entry functions of combined modules. however this part would
be deprecated by the time of second stage refactoring.

so i think this is better to do it now.

SWDEV-400726 - [LLPC][RT] Address review points

SWDEV-400726 - [LLPC][RT]Address review points