Fix acceleration structure access
Previous change for GPURT version 34 is incorrect, it only handles OpLoad of an
AS, but there are other use cases such as passing AS as function parameter,
or array of AS.

This commit makes the address space of AS pointer to be SPIRAS_Constant(4),
and lower the global variable to get.desc.ptr in llpcLowerGlobal.

The SPIRAS_Constant is needed because get.desc.ptr returns pointer with
addrspace(4).

Then any load (or GEP->load) of the AS will get its GPUVA correctly.