lgc: add initial pipeline state printer and recorder facilities
For cleaner testing of individual passes in LGC, the recorder pass is
added unconditionally to write pipeline state back into the IR if it is
present.

For debugging and testing, a dump() method and a printer pass are added
as well. They are still very incomplete, and ideally we'd address this
properly as part of some more general overhaul that allows us to write
the pipeline state into the IR in a human-readable way. For now, I've
added this to be able to add a regression test for the debug printf
lowering.