Adjust the entrypoint argument type of WorkgroupId dynamically
Current the entrypoint argument type of WorkgroupId for CS is fixed as
`<3xi32>`. The `MultiDispatchInfo` SGPR is assumed behind at the three
SGPRs for workgroupIds. It results in a mismatch issue when registers
`tgid_*_en` is optimized according to the users of workgroupId. HW will
move `MultiDispatchInfo` SGPR forward if any of workgroupId component is
not used. This change will dynamically adjust the argument type of
workgroupId and its relevant users if partial components. We add
`ShaderInput::WorkgroupId1` for using one component (`i32`) and
`ShaderInput::WorkgroupId2` for using two components (`v2i32`). The
process is handled by a new helper function `tryOptimizeWorkgroupId`.
First, check if partial components are used. Secondly, generated
corresponding `lgc.shader.input.workgroup*` call for the partial case,
relink the users to the new call and remove the old call
`lgc.shader.input.workgroup` and its users. Thirdly, add function
attributes `amdgpu-no-workgroup-id-*` for unused components. The
optimization is disabled when `m_isComputeWithCalls` is true.
NOTE: we add the optimization on `tgid_*_en` in the old new pal metadata
to verify the change.