Optimize fast link: Disable resource relocatable
Currently, relocatable shader is unsupported if there is no 1-to-1 mapping

between descriptor sets and descriptor tables, in which case we have to go back

to full pipeline, it's slower.

According to the vulkan spec, when compiling graphics pipeline library, the

resources are identified, so we don't need to use relocatable resources.