Fix mesh shader amdgpu-flat-work-group-size attribute setting
Round up the maximum flat workgroup size to a multiple of the wave size
to reflect how many lanes will actually be enabled by the initial call
to llvm.amdgcn.init.exec(-1).

Without this the maximum flat workgroup size could be set to 1, which
would cause the backend to optimize assuming that all values are
uniform. This problem is exposed by a recent upstream LLVM patch:
https://reviews.llvm.org/D151341 "AMDGPU: Special case uniformity info for single lane workgroups"