lgc: use replaceAllUsesWith() to replace function uses
The way to replace the use in Constant derived classes will break how
LLVM manage Constant. LLVM will hash all the information of a Constant
(Type, Opc, Operands), and holding the Constant information in
ConstantUniqueMap. Replacing the operand of a ConstantExpr without
updating the ConstantUniqueMap will simply invalidate the map. This
will cause a later replaceAllUsesWith() fail or a crash when destroy
the context.