Set wavefrontsize to entry-point attribute earlier
Previously, we set this sub-attribute before passing finalized IR to
backend. At that time, the wavefrontsize string is attached. However, it
is somewhat later when optimization passes do need to know this info to
make their decisions, such as removing unnecessary mbcnt.hi.

This change moves the setting of wavefrontsize earlier to avoid such
cases.