Task payload rework
In this change, we rework task payload lowering. Previously, task
payload is lowered in frontend with a special path in SPIR-V reader and
LLPC frontend lower. This is very awkward. Now, we just follow the
discussion of
https://github.com/GPUOpen-Drivers/llpc/pull/2354#pullrequestreview-1359385375.

We defer the lowering of task payload to PatchBufferOp. We just
transform the task payload buffer descriptor in MeshTaskShader lowering.
In the frontend, a virtual call is made: lgc.task.payload.ptr to represent
the getter of task payload pointer.

By doing so, many processing codes of task payload in both frontend
and in LGC could be removed to share the processing of regular buffer
operations since task payload is just a special memory buffer.