Fix problems with out-of-range shift amounts
In SPIR-V the result of a shift with an out-of-range shift amount is an
undefined value. In LLVM IR it is poison, which is more undefined. To
avoid correctness problems, mask the shift amount in IR so that it is
always in-range.

This is prompted by some recent LLVM patches which optimize poison
values more aggressively, and triggered a Vulkan CTS test failure in
dEQP-VK.graphicsfuzz.cov-fold-shift-gte32.