Freeze shift instructions
In SPIR-V the result of a shift with an out-of-range shift amount is an
undefined value. In LLVM IR it is poison, which is more undefined. To
avoid correctness problems, freeze the result of every shift just in
case it is poison. LLVM middle end optimizations should be able to
remove many of these freeze instructions, for example if it can prove
that the shift amount is in-range.

This is prompted by some recent LLVM patches which optimize poison
values more aggressively.