Handle built-ins
Handle all dialect related to built-ins.

Note that some of the built-ins are mutable during shader execution, so
instead of creating an alloca and storing initial value, map the global
to a dialect that returns a pointer.

This does not work alone, it requires upcoming commit to move
SpirvLowerRayTracing pass after SpirvLowerGlobal.