llpc: tighten up shader module creation
Propagate errors detected in SPIR-V and dump the SPIR-V binary even when
an error was detected.

This is primarily a debugging aid. When an application does not
correctly check for the presence of an extension, and we happen to have
a driver build without that extension, the compile used to crash leaving
behind a .pipe file but no .spv file (if pipeline dumping was enabled),
which thwarts the usual method of figuring out compiler crashes by
reproducing via the offline amdllpc. With this modification, the .spv
file should now be left behind as well, allowing reproduction of the
problem.