lgc: cleanup CreateLoadPushConstantsPtr
We no longer support compiles without a resource mapping, so we can assume
that if push constants are accessed, then there is a push constant resource
node.

Additionally, clean up the code after the opaque pointers transition.

We also make sure to apply instName on both code paths in the builder.