lgc/PatchEntryPointMutate: cleanup the "compute with calls" detection
We recently added support for color export shaders as tail call targets,
but those shouldn't count as compute with calls for the purposes of this
pass because their function signature does not need to be adjusted.

The fact that we don't need to adjust it correlates with the calling
convention, so use that.