Add workaround to fix GPU hang caused by incorrect scratch offset
We find a HW defect on GFX9. When there is only one user SGPR (global
table, s0), the scratch offset (s2) of PS will be incorrectly initialized
by HW. This leads to invalud scratch memory access. Note that s1 is
primitive mask and is always present according to HW spec. This defect
only occurs when scratch offset is mapped to s2.

In order to workaround this defect, we add a dummy user SGPR to avoid mapping
scratch offset to s2.

This defect is found when we made this PR:
https://github.com/GPUOpen-Drivers/llpc/pull/2624. The per-shader table
occupying s1 was removed, which leads to the CTS failure:
VK.graphicsfuzz.arr-value-set-to-arr-value-squared.