Workaround perf drop in RT demo
This commit workarounds the performance drop observed in GPSnoopy's
RayTracingInVulkan demo.

The perf drop is caused by an extra `s_waitcnt vmcnt(2)` being generated,
when the input IR has the difference on the order of `phi`s. Detail are
discussed in an internal mail thread.

Before the root cause is addressed, apply this workaround to unblock
promotion.