Add streamActive to XfbMetadata
We should check if a vertex stream is active when we encounter EmitVertex or
EndPrimitive. Currently, set its activeness of if transform feedback is
enabled or the stream itself is the rasterization stream.

In the future, when generated primitive counting is enable, we should
set the activeness of a vertex stream as well.