Add the missing symbol usage information in ShaderModuleData
We hope that all the resource usage information can be aquired from this ShaderModuleData::ShaderModuleUsage, such as input/output symbols, default uniforms, uniform blocks, shader storage blocks, images, textures, atomics and transform feedback varaibles and some metadatas, which is required when build the resource mapping node.

This change is the first step to demo how to gether input symbol info only, other resources will be added in following changes if this demo is good to go.