lgc: Add updateUserDataMap to generateEntryPointArgTys()
This flag is to control whether to update the UserDataMap.
The motivation for the change is for cps lowering, we are also calling
generateEntryPointArgTys() to build user data argument list. But we do
not want to update the userdata map for this call. Because this call
may not take account of ShaderInputs, which will generate wrong userdata
map.