Support struct-type to apply waterfall.end intrinsic
The waterfall intrinsics can only accept simple types. The backend
support has been merged. This is a front-end update. We extend
`CreateMapToInt32` to support structs and rename it to
`CreateMapToSimpleType`, `MapToInt32Func` to `MapToSimpleTypeFunc`.

The implementation of `createWaterfallLoop` then boils down to a single
call of the new `CreateMapToSimpleType` method.