Reduce instcombine iterations form 2 to 1 for non-NGG
The instcombine is run earlier a couple of times so in most cases
it is not required to run it again with 2 iterations.

Out of 10K test shaders only a handful is affected by this change,
and the differences are negligible.

This also makes it consistent with the NGG path.