Lower OpFNegate directly to fneg (#2739)
Update SPIRVReader to lower OpFNegate directly to fneg.
By doing this the code in SpirvLowerMath can be updated to remove the change to
add a canonicalize for any fneg.

The motivation for this is that InstCombine can introduce fneg
instructions (after some recent upstream llvm changes) and in this case we don't
want to blindly add a canonicalize as this invalidates e.g. xor 0x8000000 val ->
fneg val.

It makes more sense to handle the requirement to treat OpFNegate as a floating
point instruction at the point it is lowered (llvm fneg is not, hence the reason
we may need the canonicalize intrinsic).