llpc/dbginfo: correctly handle DebugFunction
Be honest about the operands of DebugFunction when reading the SPIR-V
and cleanly handle the fact that (unlike in the OpenCL debug info) the
DebugFunction instruction doesn't reference the OpFunction at all.
Setting the LLVM debug metadata can only be done after reading the
DebugFunctionDefinition instruction which links the two.