lgc: unify tail jump to color export shader
Both methods really do the same thing conceptually and in terms of data
passing (which is the crux of what matters).

The only difference is that in one case we create a `ret` instruction,
assuming that the color export shader will be pasted after the fragment
shader; and in the other case we create a `call` instruction that should
be treated as a tail call.

It turns out that in one case we forgot to set the discard state in PAL
metadata -- a good indication that unifying this a little is a good
idea.

(cherry picked from commit 0c629a271b27c9776128deb65954173a456000aa)