Fix regression of HW stream-out rework
We found a regression caused by the PR:
https://github.com/GPUOpen-Drivers/llpc/pull/2718. The CTS tests
multiple GS streams with multiple XFB buffers. The GS has such usage:

  layout (xfb_buffer = 1, xfb_stride = 64) out;
  layout (xfb_buffer = 2, xfb_stride = 64) out;
  layout (xfb_buffer = 3, xfb_stride = 64) out;

  layout (stream = 0, xfb_buffer = 1, xfb_offset = 48) out vec4 goku;
  layout (stream = 0, xfb_buffer = 1, xfb_offset = 32) out vec4 gohan;
  layout (stream = 0, xfb_buffer = 1, xfb_offset = 16) out vec4 goten;
  layout (stream = 1, xfb_buffer = 3, xfb_offset = 48) out vec4 picolo;
  layout (stream = 1, xfb_buffer = 3, xfb_offset = 32) out vec4 vegeta;
  layout (stream = 2, xfb_buffer = 2, xfb_offset = 32) out vec4 bulma;

In copy shader, we add
streamOffset0/streamOffset1/streamOffset2/streamOffset3 as 4 arguments
unconditionally but the actual SGPR initialization is controlled by the
register field SO_BASEX_EN. This means xfbBuffer[1] is mapped to
streamOffset0 because xfbBuffer[0] is not used in above case. It is the
same case for xfbBuffer[2] and xfbBuffer[3]. The mapping between
xfbBuffer[X] and streamOffset[X] must be carefully determined.

In the change, also rename some stream-out variables to match HW
semantics and remove unnecessary GFX8/GFX9 handling in stream-out
writes.