lgc: extend InputAssemlyState::enableMultiView to an enum
Vulkan view index is used to set RT layer when multiview extension is
enabled. But HLSL SV_ViewID stands for a combination of viewport index,
RT layer and view index. To support HLSL and future Vulkan
extension, this bool enableMultiView is expandede as an enum which is one of:

- Disabled
- Simple: Current Vulkan behavior, i.e. RT layer set to view index, viewport
  index set by shader
- PerView: both RT layer and viewport index set by shader (with shader output
  defaulting to 0), offset by a base that's taken from the ViewId userdata

There is also key change in RegisterMetadataBuilder because even the
viewport index is written when multiView is PerView mode, vertex reuse
can still be enabled because the viewport index is uniform.

Fix build - missed a enableMultiView