Handle missing xfb execution mode in addCallInstForXfbOutput
There is a case that the vertex outputs has xfb decoration but without
declared with xfb execution mode. As vulkan spec, these outputs will not
be captured by xfb buffer. This change will add the missing handling for
this case in `addCallInstForXfbOutput`. The check for existing xfb
outputis based on lgc.xfb.state metadata.