Fix signed zero issues of FRem
We use IR builder CreateFRem to translate OpFRem to LLVM native frem
instruction. The instruction is further lowered by backend following
such formula on our HW:

  frem(x, y) = x - y * trunc(x/y)

There is a latent issue when x=-0.0. Although SPIR-V spec says nothing
about the case when x = 0.0, C fmod function does specify the sign of
the result is the same as that of the dividend even if it is signed zero.
But when we input x=-0.0 to above formula, we finally get the addition of
(-0.0) + 0.0. The result is +0.0 returned by HW. Hence, we have to
manually check this special case when nsz fast math flag is not specified.