Fix signed zero issues of FRem
We use IR builder CreateFRem to translate OpFRem to LLVM native frem
instruction. The instruction is further lowered by backend following
such formula:

  frem(x, y) = x - y * trunc(x/y)

There is a latent issue when x=-0.0. According to SPIR-V spec, the sign
of the result of FRem is the same as the sign of the dividend. But when
we input x=-0.0 to above formula, we finally get the addition of
(-0.0) + 0.0. The result is +0.0 returned by HW. Hence, we have to
manually check this special case when nsz fast math flag is not specified.