ColorExportShader: add missing hash ingredients
The color export shader can depend on the color export state. We need to
account for that in its cache hash, otherwise we might get an incorrect
CES from the cache.

Adding all these bits to the cache hash is slightly conservative, since
not all details of the color export state are really used in generating
the CES. However, we don't have a good mechanism to automatically catch
when the generator code makes use of more of the state, so being
slightly conservative here is the right thing to do.