Disable Xfb for relocatable compilation
On gfx11, Xfb depends on the count of primitive vertex. If

UnlinkedStageVertexProcess only contains VS (no TES and GS),

the primitive type might be unknown at compiling VS time, so

we have to fall back to full pipeline.

Currently, We treat Traingle_list as the default value. Therefore,

we only disable relocatable compilation when primitive is point or

line.