llpc: Fix non-determinism in SpirvLowerRayQuery
In SpirvLowerRayQuery, we generate a unique id for each traceRay call.
This id is generated by hasing an increasing ID number, and the name
of the module.

However, the implementation was instead hashing the binary representation
of a StringRef object, thereby hashing a pointer.

This resulted in non-deterministic IR being generated.