llpc: GFX11 flat/custom parameter loads need to be strict WQM
These already use WQM annotation to ensure fetch lanes are active,
but this does not work if there are discards or divergence as the
WQM exec mask will not truly WQM (some helper invocations killed).

Hence:
- Generate strict.wqm call for flat/custom interpolation loads
- Add test coverage for this code path

Ideally this could be resolved by using the FI bit on the mov_dpp,
but there is no backend support for this currently.