[LLPC] Fully scalarize compares and selects
A while back, we switch to no longer always fully scalarizing in order
to preserve packed 16-bit operations.

This had the effect that comparisons and selects were no longer
scalarized due to limitations in LLVM's ScalarizerPass. And it seems
that there is some correctness error, most likely somewhere else in the
backend, which unfortunately hasn't been fully root caused.

In any case, there are no packed compares and (vector condition)
selects, and fully scalarizing them tends to produce better code. So
let's do that explicitly in the frontend as a quick fix.

Fixes: Red Dead Redemption with FSR 2.0
Fixes: c8f8da9cff2747 ("lgc: set scalarizer min-bits to 32")