Prepare for addition of disjoint flag from upstream llvm
Upstream llvm is introducing a new disjoint flag that makes it easier to
determine when an or instruction was originally an add - which can be useful in
some optimisations.

Since neither of the affected tests were auto-generated, I've added something
that allows for the disjoint flag to be optionally present.