Support vertex offset mode with new D3D12_VERTEX_BUFFER_VIEW (#2837)
* Support vertex offset mode with new D3D12_VERTEX_BUFFER_VIEW

In vertex offset mode. we will use D3D12_VERTEX_BUFFER_VIEW to
replace SRD, the address and size fields are already in the right
place. We only need to get stride and fill the fourth DWORD. Then we
calculate offset with stride and vertex index.

For Uber fetch shader, SRD will be updated in driver.

* Fix the lit tests