Replace getInt8PtrTy which has now been removed from upstream LLVM
This function was changed to return just getPtrTy some time ago in upstream
LLVM, but the function was left in place to assist updating.

The upstream function has now been removed with the instruction that anything
using the getInt8PtrTy builder function should switch to getPtrTy.

The code here is still required to work with typed pointers, so it is updated to
work in both cases.