Remove the unused user data EsGsLdsSize
This user data was added to keep some compatibility with old PAL. At
that time, PAL expected this user data to handle ES-GS merged shader and
NGG correctly on GFX9 and GFX10.1. But LLPC doesn't need this user data
at all since we always know ES stage when building the pipeline. And the
LDS calculation and allocation of GS stage are applied to full ES-GS
pipeline.

In this change, the user data EsGsLdsSize is not added to entry-point
argument. This SGPR is saved. Also, we don't need to tell PAL with
metadata EsGsLdsSize.